+++ packages/ui-code-editor/src/code-editor.tsx
@@ -22,9 +22,15 @@ import { bracketMatching, defaultHighlightStyle, syntaxHighlighting } from '@cod
 import { searchKeymap } from '@codemirror/search';
 import { EditorState, StateField } from '@codemirror/state';
 import { EditorView, keymap, lineNumbers } from '@codemirror/view';
-import { useEffect, useRef } from 'react';
+import * as React from 'react';
+import { useEffect, useMemo, useRef } from 'react';
 
 import styled from '@darajs/styled-components';
+
+import { getJSONExtensions } from './extensions/json';
+import { getMarkdownExtensions } from './extensions/markdown';
+import { getPythonExtensions } from './extensions/python';
+import { getSQLExtensions } from './extensions/sql';
 
 const EditorRoot = styled.div`
     overflow-y: scroll;
@@ -67,14 +73,37 @@ export interface CodeEditorProps {
      * Standard style prop
      */
     style?: React.CSSProperties;
+    /**
+     * The language to use for the code editor
+     */
+    language?: 'json' | 'python' | 'markdown' | 'sql';
 }
 
 /**
  * The CodeEditor component.
- * Currently has JSON and Python langauge support.
+ * Currently has JSON, Python, Markdown and SQL langauge support.
  */
-function CodeEditor({ initialScript, disabled, onChange, style, className }: CodeEditorProps): JSX.Element {
+function CodeEditor({ initialScript, disabled, onChange, style, className, language }: CodeEditorProps): JSX.Element {
     const editorRef = useRef();
+
+    const extensions = useMemo(() => {
+        if (!language) {
+            return [lineNumbers(), python(), json(), bracketMatching(), closeBrackets()];
+        }
+
+        switch (language) {
+            case 'json':
+                return getJSONExtensions();
+            case 'python':
+                return getPythonExtensions();
+            case 'markdown':
+                return getMarkdownExtensions();
+            case 'sql':
+                return getSQLExtensions();
+            default:
+                return [];
+        }
+    }, [language]);
 
     useEffect(() => {
         if (editorRef.current) {
@@ -98,13 +127,9 @@ function CodeEditor({ initialScript, disabled, onChange, style, className }: Cod
             const startState = EditorState.create({
                 doc: initialScript ?? '',
                 extensions: [
-                    lineNumbers(),
+                    ...extensions,
                     dispatchChanges,
                     history(),
-                    python(),
-                    json(),
-                    bracketMatching(),
-                    closeBrackets(),
                     syntaxHighlighting(defaultHighlightStyle),
                     EditorState.tabSize.of(4),
                     EditorView.editable.of(!disabled),